separate core extensions into an extra file & make ordinal a non-kernel method
